pyproject.toml
[tool.poetry]
name = "YoutubeSort"
version = "0.1.0"
description = "Sort and search youtue videos and playlists."
authors = ["Kevin Postal"]

[tool.poetry.dependencies]
python = "^3.10"
Django = "^4.0.3"
google-api-python-client = "^2.40.0"
youtube_dl = "^2021.12.17"
django-channels = "^0.7.0"
celery = "^5.2.3"
redis = "^4.1.4"
channels-redis = "^3.4.0"
psycopg2-binary = "^2.9.3"
djangorestframework = "^3.13.1"
google-auth-oauthlib = "^0.5.0"
django-celery-results = "^2.3.0"

[tool.poetry.dev-dependencies]
django-debug-toolbar = "^3.2.4"
mypy = "^0.941"
pydocstyle = "^6.1.1"
pylama = "^8.3.8"
flake8 = "^4.0.1"
bandit = "^1.7.4"
django-stubs = "^1.9.0"
djangorestframework-stubs = "^1.4.0"
celery-stubs = "^0.1.2"
types-redis = "^4.1.17"
types-psycopg2 = "^2.9.8"
wemake-python-styleguide = "^0.16.1"
yapf = "^0.32.0"
isort = "^5.10.1"
django-stubs-ext = "^0.3.1"
django-extensions = "^3.1.5"
Werkzeug = "^2.0.3"


[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
